+++ tests/testthat/test-wrap_add_root_and_add_pseudotime.R
@@ -46,7 +46,8 @@ trajectory <- wrap_data(
   divergence_regions = divergence_regions
 )
 
-test_that("Testing add_root", {
+# dynplot::plot_graph(trajectory, label_milestones = TRUE)
+test_that("Testing add_root on simple linear trajectory", {
   rooted <- add_root(trajectory, root_cell_id = "a")
 
   expect_true(rooted$root_milestone_id == "W")
@@ -58,6 +59,20 @@ test_that("Testing add_root", {
 
   expect_error(add_root(trajectory, root_cell_id = "trajectories are cool"))
   expect_error(add_root(trajectory, root_milestone_id = "trajectories are cool"))
+})
+
+
+test_that("Testing add_root on a more complex linear trajectory", {
+  trajectory2 <- wrap_data(cell_ids = "a") %>%
+    add_trajectory(
+      milestone_network = tibble(from = c("2", "1", "4", "3"), to = c("1", "4", "3", "5"), length = 1, directed = TRUE),
+      progressions = tibble(cell_id = "a", from = "2", to = "1", percentage = 0)
+    )
+
+
+  trajectory2_rooted <- add_root(trajectory2, root_milestone_id = "5")
+
+  expect_true(all(trajectory2_rooted$milestone_network$from == c("5", "3", "4", "1")))
 })
 
 
@@ -74,5 +89,5 @@ test_that("Testing add_root_using_expression",{
 test_that("Testing calculate_pseudotime", {
   trajectory <- add_root(trajectory)
   trajectory <- add_pseudotime(trajectory)
-  expect_equal(trajectory$pseudotime, c("a" = 0, "b" = 1.4, "c" = 2.6, "d" = 4.3, "e" = 4.4, "f" = 6.8))
+  expect_equal(trajectory$pseudotime, c("a" = 0.2, "b" = 1.6, "c" = 2.8, "d" = 4.5, "e" = 4.6, "f" = 7.0))
 })